fix: Coming-soon page
Fix all the issues!!!
