--- src/common/components/NoticeCard/NoticeCardError/index.ts
@@ -1,2 +1,0 @@
-export * from "./ErrorNoticeCard";
-export * from "./ErrorNoticeCardTheme";
