+++ src/common/components/NoticeCard/index.ts
@@ -1,2 +1,2 @@
 export * from "./NoticeCard";
-export * from "./NoticeCardTheme";
+export * from "./NoticeCard.theme";
