+++ src/common/components/Sidebar/Sidebar.tsx
@@ -26,16 +26,7 @@ const SIDEBAR_ITEMS = [
     href: "/bidding",
   },
 
-  // {
-  //   label: "Marketplace",
-  //   icon: <BookLineIcon size={16} />,
-  //   href: "/marketplace",
-  // },
-  // {
-  //   label: "Deals",
-  //   icon: <DealsIcon size={16} />,
-  //   href: "/deals",
-  // },
+
   
 
   // Development-only links
