+++ src/common/styles/components.scss
@@ -2,5 +2,13 @@
   .focus-ring {
     @apply focus-visible:border-primary-default focus-visible:ring-2 focus-visible:ring-primary-default/20;
   }
-
+  .break-anywhere {
+    /*
+    tailwind's `break-words` class doesn't work as expected in some cases
+    reference: https://github.com/tailwindlabs/tailwindcss/discussions/12127
+    Modzilla docs on `break-word` https://developer.mozilla.org/en-US/docs/Web/CSS/word-break#break-word
+    Playground for our usecase https://play.tailwindcss.com/x9h990fI5Y?layout=horizontal
+    */
+    overflow-wrap: anywhere;
+  }
 }
